requirements.txt.txt
langchain
langchain-community
langchain-core
langchain-groq
chromadb
sentence-transformers
python-dotenv
pypdf
docx2txt